extract main class
